requirements.txt
fastapi
uvicorn
motor
pydantic
python-multipart
python-dotenv
google-generativeai
httpx
openai
